// src/interfaces/people_types.ts
export default interface People {
    people_data: {
        "name": string
        "height": string
        "mass": string
        "hair_color": string
        "skin_color": string
        "eye_color": string
        "birth_year": string
        "gender": string
        "homeworld": string
        "films": string[]
        "species": string[]
        "vehicles": string[]
        "starships": string[]
        "url": string
        "created": string
        "edited": string
    }
}